feat(code-viewer): animate syntax highlighting
When first rendering the CodeViewer it does not perform syntax highlighting.

To optimize performance highlighting happens after rendering in a `useEffect` hook. This causes a flicker when the coloring is finally applied to the text.

This fix smoothens out the process using a short (0.3s) animation to introduce the color.